python-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
google-generativeai==0.3.1
python-dotenv==1.0.0
httpx==0.24.1